VERY wonkily fix nullable DateTime display
It seems MatBlazor (or Blazor itself?) populates the nullable
datetime with the default value (DateTime.MinValue) if it's left
empty in the form by the time we get to posting the model to the
Azure function the PurchaseDate is already set to 0001-01-01 if
you left it empty.

I presume nobody actually bought their gear in the year 1 CE, so
we just set the value to null in that case just before we send the
model to be saved.